+++ bindings/custom.cpp
@@ -11,7 +11,7 @@
 
 template<typename T>
 struct GetExtSize {
-    consteval size_t operator()() const noexcept
+    consteval std::size_t operator()() const noexcept
     {
         if constexpr (std::same_as<T, void>) {
             return 0;
@@ -23,7 +23,7 @@ struct GetExtSize {
 
 template<typename T>
 struct GetExtAlign {
-    consteval size_t operator()() const noexcept
+    consteval std::size_t operator()() const noexcept
     {
         if constexpr (std::same_as<T, void>) {
             return 0;
@@ -56,7 +56,7 @@ struct GetExtSupported {
     }
 };
 
-static inline uint8_t* wisdom_alloc(size_t size, size_t align)
+static inline uint8_t* wisdom_alloc(std::size_t size, std::size_t align)
 {
     return static_cast<uint8_t*>(ALIGNED_ALLOC(size, align));
 }
@@ -98,14 +98,14 @@ static inline decltype(auto) DX12BridgeSelector(Args&&... args)
 template<typename ExtT, typename PreType, typename EnumT>
 static inline ExtLayout<PreType, EnumT> wisdom_alloc_ext_dx12(std::span<ExtT> exts) noexcept
 {
-    size_t size = 0;
-    size_t all_exts = 0;
-    size_t first_al = 1;
+    std::size_t size = 0;
+    std::size_t all_exts = 0;
+    std::size_t first_al = 1;
     for (auto&& ext : exts) {
         if (ext.extension_id == 0) {
             continue;
         }
-        size_t align = DX12BridgeSelector<PreType, GetExtAlign>(static_cast<EnumT>(ext.extension_id));
+        std::size_t align = DX12BridgeSelector<PreType, GetExtAlign>(static_cast<EnumT>(ext.extension_id));
         if (all_exts == 0) {
             first_al = align;
         }
@@ -115,11 +115,11 @@ static inline ExtLayout<PreType, EnumT> wisdom_alloc_ext_dx12(std::span<ExtT> ex
     }
 
     // Add the size of the extension id x all_exts
-    size_t array_size = all_exts * sizeof(EnumT); // can be 0
-    size_t cap_size = sizeof(uint32_t);
+    std::size_t array_size = all_exts * sizeof(EnumT); // can be 0
+    std::size_t cap_size = sizeof(uint32_t);
 
-    size_t base_size = wis::detail::aligned_size(sizeof(PreType), alignof(uint32_t)) + cap_size + array_size;
-    size_t overall_size = wis::detail::aligned_size(base_size, first_al) + size;
+    std::size_t base_size = wis::detail::aligned_size(sizeof(PreType), alignof(uint32_t)) + cap_size + array_size;
+    std::size_t overall_size = wis::detail::aligned_size(base_size, first_al) + size;
 
     uint8_t* memory = wisdom_alloc(overall_size, alignof(PreType));
     *reinterpret_cast<uint32_t*>(memory + wis::detail::aligned_size(sizeof(PreType), alignof(uint32_t))) = all_exts;
@@ -134,7 +134,7 @@ template<typename PreType, typename EnumT>
 static inline ExtLayout<PreType, EnumT> from_stream_dx12(uint8_t* stream) noexcept
 {
     auto* pre_type = reinterpret_cast<PreType*>(stream);
-    uint8_t* ext_count = reinterpret_cast<uint8_t*>(wis::detail::aligned_size(size_t(stream) + sizeof(PreType), alignof(uint32_t)));
+    uint8_t* ext_count = reinterpret_cast<uint8_t*>(wis::detail::aligned_size(std::size_t(stream) + sizeof(PreType), alignof(uint32_t)));
 
     uint32_t count = *reinterpret_cast<uint32_t*>(ext_count);
     if (count == 0) {
@@ -143,8 +143,8 @@ static inline ExtLayout<PreType, EnumT> from_stream_dx12(uint8_t* stream) noexce
 
     auto span = std::span{ reinterpret_cast<EnumT*>(ext_count + sizeof(uint32_t)), count };
 
-    size_t align = DX12BridgeSelector<PreType, GetExtAlign>(span[0]);
-    uint8_t* ext_ptr = reinterpret_cast<uint8_t*>(wis::detail::aligned_size((size_t)span.data() + span.size_bytes(), align));
+    std::size_t align = DX12BridgeSelector<PreType, GetExtAlign>(span[0]);
+    uint8_t* ext_ptr = reinterpret_cast<uint8_t*>(wis::detail::aligned_size((std::size_t)span.data() + span.size_bytes(), align));
 
     return ExtLayout<PreType, EnumT>{
         .pre_type = pre_type,
@@ -170,13 +170,13 @@ extern "C" WisResult DX12CreateFactory(bool debug_layer, WisFactoryExtQuery* ext
         uint8_t* ext_ptr = memory.ext_memory;
         auto ext_ids = memory.exts.data();
 
-        for (size_t i = 0; i < extension_count; i++) {
+        for (std::size_t i = 0; i < extension_count; i++) {
             if (extensions[i].extension_id != 0) {
                 auto id = static_cast<wis::FactoryExtID>(extensions[i].extension_id);
                 *ext_ids = id;
 
-                size_t align = DX12FactoryExtensionBridge<GetExtAlign>(id);
-                ext_ptr = reinterpret_cast<uint8_t*>(wis::detail::aligned_size((size_t)ext_ptr, align));
+                std::size_t align = DX12FactoryExtensionBridge<GetExtAlign>(id);
+                ext_ptr = reinterpret_cast<uint8_t*>(wis::detail::aligned_size((std::size_t)ext_ptr, align));
 
                 DX12FactoryExtensionBridge<CreateExt>(id, ext_ptr);
                 *data = reinterpret_cast<wis::DX12FactoryExtension*>(ext_ptr);
@@ -198,11 +198,12 @@ extern "C" WisResult DX12CreateFactory(bool debug_layer, WisFactoryExtQuery* ext
     }
     // Check if ext is supported
     if (extensions) {
-        for (size_t i = 0; i < extension_count; i++) {
+        for (std::size_t i = 0; i < extension_count; i++) {
             if (extensions[i].extension_id != 0) {
                 bool supported = DX12FactoryExtensionBridge<GetExtSupported>(static_cast<wis::FactoryExtID>(extensions[i].extension_id), data[i]);
-                if (!supported)
+                if (!supported) {
                     extensions[i].result = nullptr;
+                }
             }
         }
     }
@@ -221,7 +222,7 @@ extern "C" void DX12FactoryDestroy(DX12Factory self)
 
     uint8_t* ext_ptr = memory.ext_memory;
     for (auto eid : memory.exts) {
-        ext_ptr = reinterpret_cast<uint8_t*>(wis::detail::aligned_size((size_t)ext_ptr, DX12FactoryExtensionBridge<GetExtAlign>(eid)));
+        ext_ptr = reinterpret_cast<uint8_t*>(wis::detail::aligned_size((std::size_t)ext_ptr, DX12FactoryExtensionBridge<GetExtAlign>(eid)));
         DX12FactoryExtensionBridge<DestroyExt>(eid, ext_ptr);
         ext_ptr += DX12FactoryExtensionBridge<GetExtSize>(eid);
     }
@@ -249,13 +250,13 @@ extern "C" WisResult DX12CreateDevice(DX12Adapter adapter, WisDeviceExtQuery* ex
         uint8_t* ext_ptr = memory.ext_memory;
         auto ext_ids = memory.exts.data();
 
-        for (size_t i = 0; i < extension_count; i++) {
+        for (std::size_t i = 0; i < extension_count; i++) {
             if (extensions[i].extension_id != 0) {
                 auto id = static_cast<wis::DeviceExtID>(extensions[i].extension_id);
                 *ext_ids = id;
 
-                size_t align = DX12DeviceExtensionBridge<GetExtAlign>(id);
-                ext_ptr = reinterpret_cast<uint8_t*>(wis::detail::aligned_size((size_t)ext_ptr, align));
+                std::size_t align = DX12DeviceExtensionBridge<GetExtAlign>(id);
+                ext_ptr = reinterpret_cast<uint8_t*>(wis::detail::aligned_size((std::size_t)ext_ptr, align));
 
                 *data = reinterpret_cast<wis::DX12DeviceExtension*>(ext_ptr);
                 DX12DeviceExtensionBridge<CreateExt>(id, *data);
@@ -277,11 +278,12 @@ extern "C" WisResult DX12CreateDevice(DX12Adapter adapter, WisDeviceExtQuery* ex
         return reinterpret_cast<WisResult&>(result);
     }
 
-    for (size_t i = 0; i < extension_count; i++) {
+    for (std::size_t i = 0; i < extension_count; i++) {
         if (extensions[i].extension_id != 0) {
             bool supported = DX12DeviceExtensionBridge<GetExtSupported>(static_cast<wis::DeviceExtID>(extensions[i].extension_id), data[i]);
-            if (!supported)
+            if (!supported) {
                 extensions[i].result = nullptr;
+            }
         }
     }
 
@@ -299,7 +301,7 @@ extern "C" void DX12DeviceDestroy(DX12Device self)
 
     uint8_t* ext_ptr = memory.ext_memory;
     for (auto eid : memory.exts) {
-        ext_ptr = reinterpret_cast<uint8_t*>(wis::detail::aligned_size((size_t)ext_ptr, DX12DeviceExtensionBridge<GetExtAlign>(eid)));
+        ext_ptr = reinterpret_cast<uint8_t*>(wis::detail::aligned_size((std::size_t)ext_ptr, DX12DeviceExtensionBridge<GetExtAlign>(eid)));
         DX12DeviceExtensionBridge<DestroyExt>(eid, ext_ptr);
         ext_ptr += DX12DeviceExtensionBridge<GetExtSize>(eid);
     }
@@ -340,14 +342,14 @@ static inline decltype(auto) VKBridgeSelector(Args&&... args)
 template<typename ExtT, typename PreType, typename EnumT>
 static inline ExtLayout<PreType, EnumT> wisdom_alloc_ext_vk(std::span<ExtT> exts) noexcept
 {
-    size_t size = 0;
-    size_t all_exts = 0;
-    size_t first_al = 1;
+    std::size_t size = 0;
+    std::size_t all_exts = 0;
+    std::size_t first_al = 1;
     for (auto&& ext : exts) {
         if (ext.extension_id == 0) {
             continue;
         }
-        size_t align = VKBridgeSelector<PreType, GetExtAlign>(static_cast<EnumT>(ext.extension_id));
+        std::size_t align = VKBridgeSelector<PreType, GetExtAlign>(static_cast<EnumT>(ext.extension_id));
         if (all_exts == 0) {
             first_al = align;
         }
@@ -357,11 +359,11 @@ static inline ExtLayout<PreType, EnumT> wisdom_alloc_ext_vk(std::span<ExtT> exts
     }
 
     // Add the size of the extension id x all_exts
-    size_t array_size = all_exts * sizeof(EnumT); // can be 0
-    size_t cap_size = sizeof(uint32_t);
+    std::size_t array_size = all_exts * sizeof(EnumT); // can be 0
+    std::size_t cap_size = sizeof(uint32_t);
 
-    size_t base_size = wis::detail::aligned_size(sizeof(PreType), alignof(uint32_t)) + cap_size + array_size;
-    size_t overall_size = wis::detail::aligned_size(base_size, first_al) + size;
+    std::size_t base_size = wis::detail::aligned_size(sizeof(PreType), alignof(uint32_t)) + cap_size + array_size;
+    std::size_t overall_size = wis::detail::aligned_size(base_size, first_al) + size;
 
     uint8_t* memory = wisdom_alloc(overall_size, alignof(PreType));
     *reinterpret_cast<uint32_t*>(memory + wis::detail::aligned_size(sizeof(PreType), alignof(uint32_t))) = all_exts;
@@ -376,7 +378,7 @@ template<typename PreType, typename EnumT>
 static inline ExtLayout<PreType, EnumT> from_stream_vk(uint8_t* stream) noexcept
 {
     auto* pre_type = reinterpret_cast<PreType*>(stream);
-    uint8_t* ext_count = reinterpret_cast<uint8_t*>(wis::detail::aligned_size(size_t(stream) + sizeof(PreType), alignof(uint32_t)));
+    uint8_t* ext_count = reinterpret_cast<uint8_t*>(wis::detail::aligned_size(std::size_t(stream) + sizeof(PreType), alignof(uint32_t)));
 
     uint32_t count = *reinterpret_cast<uint32_t*>(ext_count);
     if (count == 0) {
@@ -385,8 +387,8 @@ static inline ExtLayout<PreType, EnumT> from_stream_vk(uint8_t* stream) noexcept
 
     auto span = std::span{ reinterpret_cast<EnumT*>(ext_count + sizeof(uint32_t)), count };
 
-    size_t align = VKBridgeSelector<PreType, GetExtAlign>(span[0]);
-    uint8_t* ext_ptr = reinterpret_cast<uint8_t*>(wis::detail::aligned_size((size_t)span.data() + span.size_bytes(), align));
+    std::size_t align = VKBridgeSelector<PreType, GetExtAlign>(span[0]);
+    uint8_t* ext_ptr = reinterpret_cast<uint8_t*>(wis::detail::aligned_size((std::size_t)span.data() + span.size_bytes(), align));
 
     return ExtLayout<PreType, EnumT>{
         .pre_type = pre_type,
@@ -412,13 +414,13 @@ extern "C" WisResult VKCreateFactory(bool debug_layer, WisFactoryExtQuery* exten
         uint8_t* ext_ptr = memory.ext_memory;
         auto ext_ids = memory.exts.data();
 
-        for (size_t i = 0; i < extension_count; i++) {
+        for (std::size_t i = 0; i < extension_count; i++) {
             if (extensions[i].extension_id != 0) {
                 auto id = static_cast<wis::FactoryExtID>(extensions[i].extension_id);
                 *ext_ids = id;
 
-                size_t align = VKFactoryExtensionBridge<GetExtAlign>(id);
-                ext_ptr = reinterpret_cast<uint8_t*>(wis::detail::aligned_size((size_t)ext_ptr, align));
+                std::size_t align = VKFactoryExtensionBridge<GetExtAlign>(id);
+                ext_ptr = reinterpret_cast<uint8_t*>(wis::detail::aligned_size((std::size_t)ext_ptr, align));
 
                 VKFactoryExtensionBridge<CreateExt>(id, ext_ptr);
                 *data = reinterpret_cast<wis::VKFactoryExtension*>(ext_ptr);
@@ -440,11 +442,12 @@ extern "C" WisResult VKCreateFactory(bool debug_layer, WisFactoryExtQuery* exten
     }
     // Check if ext is supported
     if (extensions) {
-        for (size_t i = 0; i < extension_count; i++) {
+        for (std::size_t i = 0; i < extension_count; i++) {
             if (extensions[i].extension_id != 0) {
                 bool supported = VKFactoryExtensionBridge<GetExtSupported>(static_cast<wis::FactoryExtID>(extensions[i].extension_id), data[i]);
-                if (!supported)
+                if (!supported) {
                     extensions[i].result = nullptr;
+                }
             }
         }
     }
@@ -463,7 +466,7 @@ extern "C" void VKFactoryDestroy(VKFactory self)
 
     uint8_t* ext_ptr = memory.ext_memory;
     for (auto eid : memory.exts) {
-        ext_ptr = reinterpret_cast<uint8_t*>(wis::detail::aligned_size((size_t)ext_ptr, VKFactoryExtensionBridge<GetExtAlign>(eid)));
+        ext_ptr = reinterpret_cast<uint8_t*>(wis::detail::aligned_size((std::size_t)ext_ptr, VKFactoryExtensionBridge<GetExtAlign>(eid)));
         VKFactoryExtensionBridge<DestroyExt>(eid, ext_ptr);
         ext_ptr += VKFactoryExtensionBridge<GetExtSize>(eid);
     }
@@ -489,13 +492,13 @@ extern "C" WisResult VKCreateDevice(VKAdapter adapter, WisDeviceExtQuery* extens
         uint8_t* ext_ptr = memory.ext_memory;
         auto ext_ids = memory.exts.data();
 
-        for (size_t i = 0; i < extension_count; i++) {
+        for (std::size_t i = 0; i < extension_count; i++) {
             if (extensions[i].extension_id != 0) {
                 auto id = static_cast<wis::DeviceExtID>(extensions[i].extension_id);
                 *ext_ids = id;
 
-                size_t align = VKDeviceExtensionBridge<GetExtAlign>(id);
-                ext_ptr = reinterpret_cast<uint8_t*>(wis::detail::aligned_size((size_t)ext_ptr, align));
+                std::size_t align = VKDeviceExtensionBridge<GetExtAlign>(id);
+                ext_ptr = reinterpret_cast<uint8_t*>(wis::detail::aligned_size((std::size_t)ext_ptr, align));
 
                 *data = reinterpret_cast<wis::VKDeviceExtension*>(ext_ptr);
                 VKDeviceExtensionBridge<CreateExt>(id, *data);
@@ -517,11 +520,12 @@ extern "C" WisResult VKCreateDevice(VKAdapter adapter, WisDeviceExtQuery* extens
         return reinterpret_cast<WisResult&>(result);
     }
 
-    for (size_t i = 0; i < extension_count; i++) {
+    for (std::size_t i = 0; i < extension_count; i++) {
         if (extensions[i].extension_id != 0) {
             bool supported = VKDeviceExtensionBridge<GetExtSupported>(static_cast<wis::DeviceExtID>(extensions[i].extension_id), data[i]);
-            if (!supported)
+            if (!supported) {
                 extensions[i].result = nullptr;
+            }
         }
     }
 
@@ -539,7 +543,7 @@ extern "C" void VKDeviceDestroy(VKDevice self)
 
     uint8_t* ext_ptr = memory.ext_memory;
     for (auto eid : memory.exts) {
-        ext_ptr = reinterpret_cast<uint8_t*>(wis::detail::aligned_size((size_t)ext_ptr, VKDeviceExtensionBridge<GetExtAlign>(eid)));
+        ext_ptr = reinterpret_cast<uint8_t*>(wis::detail::aligned_size((std::size_t)ext_ptr, VKDeviceExtensionBridge<GetExtAlign>(eid)));
         VKDeviceExtensionBridge<DestroyExt>(eid, ext_ptr);
         ext_ptr += VKDeviceExtensionBridge<GetExtSize>(eid);
     }