+++ wisdom/shared.cmake
@@ -43,10 +43,9 @@ endif()
 
 set_target_properties(wisdom-shared PROPERTIES CXX_STANDARD 20)
 target_compile_features(wisdom-shared INTERFACE cxx_std_20)
-if (MSVC)
-    # Add the /Zc:__cplusplus flag
-    target_compile_options(wisdom-shared INTERFACE "/Zc:__cplusplus")
-endif()
+target_compile_options(wisdom-shared INTERFACE
+  $<$<CXX_COMPILER_ID:MSVC>:/Zc:__cplusplus>
+)
 
 target_compile_definitions(
   wisdom-shared